Use fixture characters for testing
