+++ src/character/__tests__/getPhrases.test.ts
@@ -1,7 +1,10 @@
 import { getPhrases } from "../getPhrases";
 import { King } from "../../king/types";
 import getKing from "../../king/factory/getKing";
-import { fixtureCharacters } from "../fixtures";
+import { mariaSarmiento } from "../../king/fixtures";
+import { carbassot } from "../../fighter/fixtures";
+import { bradPitt } from "../../squire/fixtures";
+import { paquitoChocolatero } from "../../adviser/fixtures";
 
 describe("Given the getPhrases funtion", () => {
   describe("When it receives 3 Maria Sarmiento", () => {
@@ -64,10 +67,17 @@ describe("Given the getPhrases funtion", () => {
   });
 
   describe("When it receives María Sarmiento, Carbassot Carbassa, Paquito Chocolatero and Brad Pitt", () => {
+    const characters = [
+      mariaSarmiento,
+      carbassot,
+      paquitoChocolatero,
+      bradPitt,
+    ];
+
     test("Then it should return 'You are all going to die' in the first position", () => {
       const expectedPhrase = "You are all going to die";
 
-      const characterPhrases = getPhrases(fixtureCharacters);
+      const characterPhrases = getPhrases(characters);
       const phrase = characterPhrases[0];
 
       expect(phrase).toBe(expectedPhrase);
@@ -76,7 +86,7 @@ describe("Given the getPhrases funtion", () => {
     test("Then it should return 'First I punch, then I ask' in the second position", () => {
       const expectedPhrase = "First I punch, then I ask";
 
-      const characterPhrases = getPhrases(fixtureCharacters);
+      const characterPhrases = getPhrases(characters);
       const phrase = characterPhrases[1];
 
       expect(phrase).toBe(expectedPhrase);
@@ -86,7 +96,7 @@ describe("Given the getPhrases funtion", () => {
       const expectedPhrase =
         "I don't know why, but I think I'm going to die soon";
 
-      const characterPhrases = getPhrases(fixtureCharacters);
+      const characterPhrases = getPhrases(characters);
       const phrase = characterPhrases[2];
 
       expect(phrase).toBe(expectedPhrase);
@@ -95,7 +105,7 @@ describe("Given the getPhrases funtion", () => {
     test("Then it should return 'I'm a loser' in the fourth position", () => {
       const expectedPhrase = "I'm a loser";
 
-      const characterPhrases = getPhrases(fixtureCharacters);
+      const characterPhrases = getPhrases(characters);
       const phrase = characterPhrases[3];
 
       expect(phrase).toBe(expectedPhrase);
